add github oauth to sveltekit test
