+++ test-apps/sveltekit/src/routes/+layout.svelte
@@ -29,7 +29,6 @@
 	:global(input:not(.checkbox)) {
 		@apply border appearance-none outline-none my-1 rounded p-1 w-full;
 	}
-	:global(.github),
 	:global(.button),
 	:global(button) {
 		@apply bg-black text-white my-2 rounded py-1 cursor-pointer w-full text-center hover:opacity-80;
