+++ test-apps/sveltekit/src/routes/+page.server.ts
@@ -1,0 +1,8 @@
+import { redirect } from '@sveltejs/kit';
+import type { PageServerLoad } from './$types';
+
+export const load: PageServerLoad = async ({ locals }) => {
+	const session = await locals.getSession();
+	if (!session) throw redirect(302, '/login');
+	return {};
+};
