+++ test-apps/sveltekit/src/routes/+page.svelte
@@ -4,20 +4,14 @@
 	const user = getUser();
 </script>
 
-{#if $user}
-	<p>This page is protected and can only be accessed by authenticated users.</p>
-	<pre class="code">
+<p>This page is protected and can only be accessed by authenticated users.</p>
+<pre class="code">
 {JSON.stringify($user, null, 2)}
 </pre>
 
-	<button
-		on:click={async () => {
-			await signOut();
-			invalidateAll();
-		}}>Sign out</button
-	>
-{:else}
-	<h2>Sign in</h2>
-	<a href="/login" class="button">Username and password</a>
-	<a href="/api/oauth?provider=github" class="button">Github</a>
-{/if}
+<button
+	on:click={async () => {
+		await signOut();
+		invalidateAll();
+	}}>Sign out</button
+>
