+++ test-apps/sveltekit/src/routes/login/+layout.server.ts
@@ -1,0 +1,8 @@
+import { redirect } from '@sveltejs/kit';
+import type { LayoutServerLoad } from './$types';
+
+export const load: LayoutServerLoad = async ({ locals }) => {
+	const session = await locals.getSession();
+	if (session) throw redirect(302, '/');
+	return {};
+};
