+++ test-apps/sveltekit/src/routes/login/+page.svelte
@@ -1,30 +1,3 @@
-<script lang="ts">
-	import { enhance } from '$app/forms';
-	import Github from '$lib/components/Github.svelte';
-
-	export let form: { message?: string };
-</script>
-
 <h2>Sign in</h2>
-<form
-	method="post"
-	use:enhance={({ data, cancel }) => {
-		form = {};
-		const username = data.get('username')?.toString() || '';
-		const password = data.get('password')?.toString() || '';
-		if (!username || !password) {
-			form.message = 'Invalid input';
-			cancel();
-		}
-	}}
->
-	<label for="username">username</label><br />
-	<input id="username" name="username" /><br />
-	<label for="password">password</label><br />
-	<input type="password" id="password" name="password" /><br />
-	<input type="submit" value="Continue" class="button" />
-</form>
-{#if form?.message}
-	<p class="error">{form.message || ''}</p>
-{/if}
-<a href="/signup" class="link">Create a new account</a>
+<a href="/login/email" class="button">Username and password</a>
+<a href="/api/oauth?provider=github" class="button">Github</a>
