+++ test-apps/sveltekit/src/routes/login/email/+page.server.ts
@@ -33,12 +33,3 @@ export const actions: Actions = {
 		}
 	}
 };
-
-import { redirect } from '@sveltejs/kit';
-import type { PageServerLoad } from './$types';
-
-export const load: PageServerLoad = async ({ locals }) => {
-	const session = await locals.getSession();
-	if (session) throw redirect(302, '/');
-	return {};
-};
